+++ src/parser/block/ulist.test.ts
@@ -52,6 +52,10 @@ describe('Unit: parser/ulist', () => {
       assert.deepStrictEqual(inspect(parser('- 1\n - 2\n  - 3')), [['<ul><li>1<ul><li>2<ul><li>3</li></ul></li></ul></li></ul>'], '']);
       assert.deepStrictEqual(inspect(parser('- 1\n - 2\n- 3')), [['<ul><li>1<ul><li>2</li></ul></li><li>3</li></ul>'], '']);
       assert.deepStrictEqual(inspect(parser('- 1\n + 2')), [['<ul><li>1<ul><li>2</li></ul></li></ul>'], '']);
+      assert.deepStrictEqual(inspect(parser('- 1\n 1')), [['<ul><li>1<ol start="1" type="1"><li></li></ol></li></ul>'], '']);
+      assert.deepStrictEqual(inspect(parser('- 1\n 1.')), [['<ul><li>1<ol start="1" type="1"><li></li></ol></li></ul>'], '']);
+      assert.deepStrictEqual(inspect(parser('- 1\n 1. ')), [['<ul><li>1<ol start="1" type="1"><li></li></ol></li></ul>'], '']);
+      assert.deepStrictEqual(inspect(parser('- 1\n 1. 2')), [['<ul><li>1<ol start="1" type="1"><li>2</li></ol></li></ul>'], '']);
     });
 
     it('checkbox', () => {